+++ src/app/routes/broker-ui/broker-ui-routing.module.ts
@@ -3,6 +3,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {BrokerUiComponent} from './broker-ui.component';
 import {CatalogPageComponent} from './catalog-page/catalog-page/catalog-page.component';
 import {ConnectorPageComponent} from './connector-page/connector-page/connector-page.component';
+import {LegalNoticePageComponent} from './legal-notice-page/legal-notice-page/legal-notice-page.component';
 
 export const routes: Routes = [
   {
@@ -21,6 +22,15 @@ export const routes: Routes = [
     data: {
       title: 'Connectors',
       icon: 'link',
+      hideInNav: true,
+    },
+  },
+  {
+    path: 'legal-notice',
+    component: LegalNoticePageComponent,
+    data: {
+      title: 'Legal Notice',
+      icon: 'info',
       hideInNav: true,
     },
   },