+++ src/app/routes/broker-ui/legal-notice-page/legal-notice-page.module.ts
@@ -1,0 +1,15 @@
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+import {LegalNoticePageComponent} from './legal-notice-page/legal-notice-page.component';
+
+@NgModule({
+  imports: [
+    // Angular
+    CommonModule,
+    RouterModule,
+  ],
+  declarations: [LegalNoticePageComponent],
+  exports: [LegalNoticePageComponent],
+})
+export class LegalNoticePageModule {}